+++ Net.Utils.Tests/PingEntityTests.cs
@@ -56,14 +56,17 @@ namespace Net.Utils.Tests
                 {
                     foreach (var timeoutTask in EnumValues.GetTimeoutMs())
                     {
-                        Assert.DoesNotThrow(() =>
+                        foreach (var useStopWatch in EnumValues.GetBool())
                         {
-                            _ = new PingEntity(timeoutPing, timeoutTask, isRepeat);
-                        });
-                        Assert.DoesNotThrowAsync(async () => await Task.Run(() =>
-                        {
-                            _ = new PingEntity(timeoutPing, timeoutTask, isRepeat);
-                        }));
+                            Assert.DoesNotThrow(() =>
+                            {
+                                _ = new PingEntity(timeoutPing, timeoutTask, false, useStopWatch);
+                            });
+                            Assert.DoesNotThrowAsync(async () => await Task.Run(() =>
+                            {
+                                _ = new PingEntity(timeoutPing, timeoutTask, false, useStopWatch);
+                            }));
+                        }
                     }
                 }
             }
@@ -72,7 +75,7 @@ namespace Net.Utils.Tests
         }
 
         [Test]
-        public void OpenTask_DoesNotThrow()
+        public void Open_DoesNotThrow()
         {
             Utils.MethodStart();
 
@@ -82,26 +85,73 @@ namespace Net.Utils.Tests
                 {
                     Assert.DoesNotThrow(() =>
                     {
-                        foreach (var isTaskWait in EnumValues.GetBool())
+                        foreach (var useStopWatch in EnumValues.GetBool())
                         {
-                            var ping = new PingEntity(timeoutPing: timeoutPing, timeoutTask: timeoutTask, isRepeat: false);
+                            var ping = new PingEntity(timeoutPing: timeoutPing, timeoutTask: timeoutTask, useRepeat: false, useStopWatch: useStopWatch);
                             ping.Hosts.Add("localhost");
                             ping.Hosts.Add("google.com");
-                            ping.OpenTask(isTaskWait);
-                            if (isTaskWait)
-                                TestContext.WriteLine($@"{ping.Status}");
+                            ping.Hosts.Add("google-fake.com");
+                            ping.Open();
+                            TestContext.WriteLine($@"{ping.Status}");
                         }
                     });
                     Assert.DoesNotThrowAsync(async () => await Task.Run(() =>
                     {
-                        foreach (var isTaskWait in EnumValues.GetBool())
+                        foreach (var useStopWatch in EnumValues.GetBool())
                         {
-                            var ping = new PingEntity(timeoutPing: timeoutPing, timeoutTask: timeoutTask, isRepeat: false);
+                            var ping = new PingEntity(timeoutPing: timeoutPing, timeoutTask: timeoutTask, useRepeat: false, useStopWatch: useStopWatch);
                             ping.Hosts.Add("localhost");
                             ping.Hosts.Add("google.com");
-                            ping.OpenTask(isTaskWait);
-                            if (isTaskWait)
-                                TestContext.WriteLine($@"{ping.Status}");
+                            ping.Hosts.Add("google-fake.com");
+                            ping.Open();
+                            TestContext.WriteLine($@"{ping.Status}");
+                        }
+                    }));
+                }
+            }
+
+            Utils.MethodComplete();
+        }
+
+        [Test]
+        public void OpenAsync_DoesNotThrow()
+        {
+            Utils.MethodStart();
+
+            foreach (var timeoutPing in EnumValues.GetTimeoutMs())
+            {
+                foreach (var timeoutTask in EnumValues.GetTimeoutMs())
+                {
+                    Assert.DoesNotThrow(() =>
+                    {
+                        foreach (var useStopWatch in EnumValues.GetBool())
+                        {
+                            var ping = new PingEntity(timeoutPing: timeoutPing, timeoutTask: timeoutTask, useRepeat: false, useStopWatch: useStopWatch);
+                            ping.Hosts.Add("localhost");
+                            ping.Hosts.Add("google.com");
+                            ping.Hosts.Add("google-fake.com");
+                            var task = Task.Run(async () =>
+                            {
+                                await ping.OpenAsync().ConfigureAwait(true);
+                            });
+                            task.Wait();
+                            TestContext.WriteLine($@"{ping.Status}");
+                        }
+                    });
+                    Assert.DoesNotThrowAsync(async () => await Task.Run(() =>
+                    {
+                        foreach (var useStopWatch in EnumValues.GetBool())
+                        {
+                            var ping = new PingEntity(timeoutPing: timeoutPing, timeoutTask: timeoutTask, useRepeat: false, useStopWatch: useStopWatch);
+                            ping.Hosts.Add("localhost");
+                            ping.Hosts.Add("google.com");
+                            ping.Hosts.Add("google-fake.com");
+                            var task = Task.Run(async () =>
+                            {
+                                await ping.OpenAsync().ConfigureAwait(true);
+                            });
+                            task.Wait();
+                            TestContext.WriteLine($@"{ping.Status}");
                         }
                     }));
                 }